# docs/swagger.json
{"swagger":"2.0","info":{"description":"This is ketabdoozak api doc","title":"Ketabdoozak","termsOfService":"http://swagger.io/terms/","contact":{},"version":"1.0"},"paths":{"/filemanager/files/{filename}":{"get":{"security":[{"OAuth2Implicit":[]}],"tags":["filemanager"],"summary":"Download a file","parameters":[{"type":"string","description":"Filename","name":"filename","in":"path","required":true}],"responses":{"307":{"description":"Temporary Redirect"}}},"delete":{"security":[{"OAuth2Implicit":[]}],"tags":["filemanager"],"summary":"Delete a file","parameters":[{"type":"string","description":"Filename","name":"filename","in":"path","required":true}],"responses":{"204":{"description":"No Content"}}}},"/filemanager/upload":{"post":{"security":[{"OAuth2Implicit":[]}],"consumes":["multipart/form-data"],"tags":["filemanager"],"summary":"Upload a file","parameters":[{"type":"file","description":"File key","name":"file","in":"formData","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/filemanager.File"}}}}},"/listing/items":{"post":{"security":[{"OAuth2Implicit":[]}],"tags":["listing"],"summary":"Create new item","parameters":[{"description":"Request body","name":"req","in":"body","required":true,"schema":{"$ref":"#/definitions/listing.CreateItemRequest"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/listing.Item"}}}}},"/listing/items/{itemId}":{"get":{"security":[{"OAuth2Implicit":[]}],"tags":["listing"],"summary":"Get item","parameters":[{"type":"string","description":"Item id","name":"itemId","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/listing.Item"}}}},"put":{"security":[{"OAuth2Implicit":[]}],"tags":["listing"],"summary":"Update item","parameters":[{"type":"string","description":"Item id","name":"itemId","in":"path","required":true},{"description":"Request body","name":"req","in":"body","required":true,"schema":{"$ref":"#/definitions/listing.UpdateItemRequest"}}],"responses":{"204":{"description":"No Content"}}},"delete":{"security":[{"OAuth2Implicit":[]}],"tags":["listing"],"summary":"Delete item","parameters":[{"type":"string","description":"Item id","name":"itemId","in":"path","required":true}],"responses":{"204":{"description":"No Content"}}}},"/listing/items/{itemId}/archive":{"put":{"security":[{"OAuth2Implicit":[]}],"tags":["listing"],"summary":"Publish item","parameters":[{"type":"string","description":"Item id","name":"itemId","in":"path","required":true}],"responses":{"204":{"description":"No Content"}}}},"/listing/items/{itemId}/publish":{"put":{"security":[{"OAuth2Implicit":[]}],"tags":["listing"],"summary":"Publish item","parameters":[{"type":"string","description":"Item id","name":"itemId","in":"path","required":true}],"responses":{"204":{"description":"No Content"}}}},"/listing/items/{itemId}/send-for-publish":{"put":{"security":[{"OAuth2Implicit":[]}],"tags":["listing"],"summary":"Send item for publish","parameters":[{"type":"string","description":"Item id","name":"itemId","in":"path","required":true}],"responses":{"204":{"description":"No Content"}}}},"/listing/locations":{"get":{"tags":["listing"],"summary":"List locations","responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/listing.ListLocationsResponse"}}}}},"/listing/my-items":{"get":{"security":[{"OAuth2Implicit":[]}],"tags":["listing"],"summary":"List my items","responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/listing.ListItemsResponse"}}}}},"/listing/published-items":{"get":{"tags":["listing"],"summary":"List published items","responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/listing.ListItemsResponse"}}}}},"/listing/published-items/{itemId}":{"get":{"tags":["listing"],"summary":"Get published item","parameters":[{"type":"string","description":"Item id","name":"itemId","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/listing.Item"}}}}}},"definitions":{"filemanager.File":{"type":"object","properties":{"contentType":{"type":"string"},"createdAt":{"type":"string"},"filename":{"type":"string"},"size":{"type":"integer"},"updatedAt":{"type":"string"}}},"listing.CreateItemRequest":{"type":"object","properties":{"asDraft":{"type":"boolean"},"contactInfo":{"type":"array","items":{"$ref":"#/definitions/listing.ItemContactInfo"}},"description":{"type":"string"},"locationId":{"type":"string"},"ownerName":{"type":"string"},"thumbnailUrl":{"type":"string"},"title":{"type":"string"},"types":{"type":"array","items":{"$ref":"#/definitions/listing.ItemType"}}}},"listing.Item":{"type":"object","properties":{"contactInfo":{"type":"array","items":{"$ref":"#/definitions/listing.ItemContactInfo"}},"createdAt":{"type":"string"},"description":{"type":"string"},"id":{"type":"string"},"lent":{"type":"boolean"},"locationId":{"type":"string"},"locationTitle":{"type":"string"},"ownerId":{"type":"string"},"ownerName":{"type":"string"},"publishedAt":{"type":"string"},"status":{"$ref":"#/definitions/listing.ItemStatus"},"thumbnailUrl":{"type":"string"},"title":{"type":"string"},"types":{"type":"array","items":{"$ref":"#/definitions/listing.ItemType"}},"updatedAt":{"type":"string"}}},"listing.ItemContactInfo":{"type":"object","properties":{"type":{"$ref":"#/definitions/listing.ItemContactInfoType"},"value":{"type":"string"}}},"listing.ItemContactInfoType":{"type":"string","enum":["phoneNumber","sms","telegram","whatsapp"],"x-enum-varnames":["ItemContactInfoTypePhoneNumber","ItemContactInfoTypeSMS","ItemContactInfoTypeTelegram","ItemContactInfoTypeWhatsapp"]},"listing.ItemStatus":{"type":"string","enum":["draft","pendingReview","published","rejected","expired","archived","deleted"],"x-enum-varnames":["ItemStatusDraft","ItemStatusPendingReview","ItemStatusPublished","ItemStatusRejected","ItemStatusExpired","ItemStatusArchived","ItemStatusDeleted"]},"listing.ItemType":{"type":"string","enum":["donate","exchange","lend","sell"],"x-enum-varnames":["ItemTypeDonate","ItemTypeExchange","ItemTypeLend","ItemTypeSell"]},"listing.ListItemsResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/definitions/listing.Item"}}}},"listing.ListLocationsResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/definitions/listing.Location"}}}},"listing.Location":{"type":"object","properties":{"id":{"type":"string"},"parentID":{"type":"string"},"title":{"type":"string"}}},"listing.UpdateItemRequest":{"type":"object","properties":{"asDraft":{"type":"boolean"},"contactInfo":{"type":"array","items":{"$ref":"#/definitions/listing.ItemContactInfo"}},"description":{"type":"string"},"lent":{"type":"boolean"},"locationId":{"type":"string"},"ownerName":{"type":"string"},"thumbnailUrl":{"type":"string"},"title":{"type":"string"},"types":{"type":"array","items":{"$ref":"#/definitions/listing.ItemType"}}}}},"securityDefinitions":{"OAuth2Implicit":{"type":"oauth2","flow":"implicit","authorizationUrl":"https://auth.applicaset.com/realms/ketabdoozak/protocol/openid-connect/auth"}}}